Rework how mic works (fixes #76, fixes #63)
Previously, the mic gain node had both the calibration value as well as the volume control (either the calibration value or zero). This made it impossible to recalibrate the mic while on air, and also broke the calibration slider.

This commit moves it into two separate gain nodes, "calibration gain" and "mix gain", with the analyser node in the middle. It also means that the mic is now calibrated before compression.